+++ init.php
@@ -3,11 +3,11 @@
 /**
  * Plugin Name: Resurs Bank Payments for WooCommerce
  * Description: Connect Resurs Bank as WooCommerce payment gateway.
- * WC Tested up to: 9.2.3
+ * WC Tested up to: 9.4.2
  * WC requires at least: 7.6.0
  * Plugin requires ecom: master
  * Requires PHP: 8.1
- * Version: 1.0.50
+ * Version: 1.0.56
  * Author: Resurs Bank AB
  * Author URI: https://developers.resurs.com/
  * Plugin URI: https://developers.resurs.com/platform-plugins/woocommerce/resurs-merchant-api-2.0-for-woocommerce/
@@ -18,10 +18,9 @@
  * @noinspection PhpDefineCanBeReplacedWithConstInspection
  */
 
-// @todo This is wordpress. SideEffects can not be handled properly while the init looks like this.
-// @todo Consider honoring this in the future another way.
+// Welcome to WordPress. SideEffects can not be handled properly while the init looks like this.
+// Consider honoring this in the future another way.
 // phpcs:disable PSR1.Files.SideEffects
-
 
 declare(strict_types=1);
 
@@ -38,7 +37,22 @@ if (!defined(constant_name: 'ABSPATH')) {
     exit;
 }
 
-// Name of plugin directory; "resursbank"
+require_once __DIR__ . '/src/Autoloader/requirements.php';
+// Do not remove this! It ensures the plugin does not crash the entire site if ecom2
+// has not been checked out properly. This issue typically occurs during a manual
+// checkout when ecom2 is missing. We cannot move this into a class, as the autoload
+// process will fail if ecom2 is unavailable.
+if (!file_exists(filename: __DIR__ . '/lib/ecom/composer.json')) {
+    resursBankHasNoEcom();
+    return;
+}
+if (PHP_VERSION_ID < 80100) {
+    resursBankHasOldPhp();
+    return;
+}
+
+
+// Name of plugin directory; normally the slug name.
 define(
     constant_name: 'RESURSBANK_MODULE_DIR_NAME',
     value: substr(
@@ -47,7 +61,7 @@ define(
     )
 );
 
-// Absolute path to plugin directory; "/var/www/html/wp-content/plugins/resursbank"
+// Absolute path to plugin directory; "/var/www/html/wp-content/plugins/<the-slug-name>>"
 define(
 	constant_name: 'RESURSBANK_MODULE_DIR_PATH',
 	value: plugin_dir_path(file: __FILE__)
@@ -89,6 +103,7 @@ add_action(hook_name: 'plugins_loaded', callback: static function (): void {
                 return;
             }
 
+            /** @noinspection PhpArgumentWithoutNamedIdentifierInspection */
             FeaturesUtil::declare_compatibility(
                 'custom_order_tables',
                 __FILE__,